feat(core): resolve symlinks on file write
Instead of using the passed filename as the rename target, use
Path.resolve for resolving any symlinks.

This allows to use symlinks for the nvchecker database files that point
somewhere else. Before this commit nvchecker simply replaced the symlink
with an actual file, now we resolve the targets first and only replace
the actual file that is pointed to.

Co-authored-by: Andreas 'Segaja' Schleifer <webmaster@segaja.de>
Signed-off-by: Levente Polyak <levente@leventepolyak.net>